backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
opencv-python
mediapipe
torch  
transformers  
huggingface-hub
Pillow
tensorflow  